Implement SinglyLinkedList class and methods, and Node inner class
The build.gradle file had to be amended to include the main class name.
This commit is a bit late in the process. As of this commit the
following methods have been implemented: first, last, addLast, addFirst,
removeFirst, get, size, and isEmpty. I have decided to implement the
SinglyLinkedList structure using nodes. These are implemented using the
inner class Node.