Fully implement remove method
I wrote unit tests to ensure that the head and tail of the list are
updated in the event that the removed node is the first or last element
of the list. The method removeLast() should cover the latter case,
however there is no guarantee that it will be used to remove the last
element.